Update to 6.1.1 API, gauges only report when changed

+++ src/main/java/org/avaje/metric/core/DefaultValueStatistics.java
@@ -35,7 +35,7 @@ public class DefaultValueStatistics implements ValueStatistics {
     this.total = total;
     // collection is racy so sanitize the max value if it has not been set
     // this most likely would happen when count = 1 so max = mean
-    this.max = max != Long.MIN_VALUE ? max : (count < 1 ? 0 : Math.round((double)(total / count)));
+    this.max = max != Long.MIN_VALUE ? max : (count < 1 ? 0 : Math.round(total / count));
   }
   
   public String toString() {
